# server/common/api.yml
swagger: '2.0'
info:
  version: 1.0.0
  title: petApi
  description: Sample API for Getty
basePath: /api/v1
...
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  /pet/{id}:
    get:
      tags:
        - PetManagement
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the pet to retrieve
          type: string
      responses:
        200:
          description: Return a pet with the specified id
        404:
          description: Pet not found
...